fix(app-delivery): could not use PipelineDeployStackAction more than once in a Stage (#8217)
The PipelineDeployStackAction class hard-coded the names of the CloudFormation
actions it created for preparing and executing the ChangeSet.
Which meant this composite action could never be used more than once in the same stage
(as that would result in errors saying the action names were duplicated).
Allow overriding the names of the actions to make it possible to be used
more than once in the same CodePipeline stage.

Fixes #3984
Fixes #8183

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*