fix(codepipeline): correctly handle CODEBUILD_CLONE_REF in BitBucket source (#7107)
As it turns out, when using the `OutputArtifactFormat` equal to `CODEBUILD_CLONE_REF` in the BitBucket source action requires the subsequent CodeBuild project to have UseConnection permissions. Use CodePipeline's `Artifact` class to transfer that information between
the source and build actions, by adding the capability to store arbitrary metadata inside the `Artifact` class.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*