feat(cognito): sign in url for a UserPoolDomain (#8155)
Compute the sign in URL from a user pool domain, given a client.

The previous defaults on the UserPoolClient created one successfully but
was unusable since all of the features were turned off.
The defaults have been changed now so that the client created with the
defaults works out of the box.

BREAKING CHANGE: OAuth flows `authorizationCodeGrant` and
`implicitCodeGrant` in `UserPoolClient` are enabled by default.
* **cognito:** `callbackUrl` property in `UserPoolClient` is now
optional and has a default.
* **cognito:** All OAuth scopes in a `UserPoolClient` are now enabled
by default.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*