fix(lambda): `SingletonFunction.grantInvoke()` API fails with error 'No child with id'
Updates to the Grant API[1] introduced the need to return the statement
that was added as a result of the grant operation.

A corresponding change[2] was applied to lambda module's `FunctionBase`
class with the intention to apply this across all constructs that are
variants of `Function`.

However, the `SingletonFunction` construct behaves differently in how it
modifies the construct tree. Specifically, it contains no child node but
instead manipulates a node that is a direct child of the `Stack` node.
For this reason, `this.node.findChild()` API does not return the
expected underlying node.

The fix here is to allow such special inheritors of `FunctionBase` to
override where the child node is to be found, via an internal method
called `_functionNode()`.

[1]: https://github.com/aws/aws-cdk/commit/1819a6b5920bb22a60d09de870ea625455b90395
[2]: https://github.com/aws/aws-cdk/commit/1819a6b5920bb22a60d09de870ea625455b90395#diff-73cb0d8933b87960893373bd263924e2