feat(ec2): allow filtering ip exhausted subnets
When providing SubnetSelection it's useful to ignore Subnets in a
selected subnet group that have 0 available IPs, since we generally
wouldn't want to try to deploy to those subnets.

For example if we have a subnet group with 4 subnets, 3 of which have
sufficient IP space: this is the case in our organization.

closes #8310