fix(rds): cannot delete a stack with DbCluster set to 'Retain' (#8110)
When the DatabaseCluster has its deletion policy set to 'Retain',
an attempt to delete the stack containing it fails,
as the DbSubnetGroup cannot be removed if it still points to an existing Cluster.
To fix that, set the retention policy of DbSubnetGroup to 'Retain'
if it is 'Retain' on the DatabaseCluster.

Fixes #5282

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*