fix(codepipeline): allow multiple CodeCommit source actions using events (#8018)
There are use-cases when you want to add the same CodeCommit
repository to a CodePipeline multiple times, with different branches.
This wouldn't work when using CloudWatch Events to trigger the pipeline,
as the ID of the generated Event only used the pipeline ID for uniqueness.
Change it to also use the branch name when generating the Event ID
(which cannot be empty, as it turns out, so validate that as well).

Fixes #7802

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*