chore(cli): fix security related changes integ test (#8274)
The test was passing because (1) the stack contained an incorrect service principal
and (2) `--require-approval` was set to `never` by default. This means that the stack
was actually deployed but failed, making the test pass.

Corrected the service principal, added an expectation to ensure that the stack
did not deploy and removed the `--require-approval` CLI option during this test.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*