chore(cli): fix integ test for IAM diff (#8316)
following on from #8274 where a broken integ test that never ran was fixed.
A different test for the IAM diff was verifying previously incorrect service
principal and we missed updating the expectation.

This fixes up the expectations to align to the changes made in #8274 and
uses the corrected service principal (ec2.amazonaws.com).

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*