chore(bootstrap): split file/image publishing roles (#8319)
For security purposes, we decided that it would be lower risk to assume a different role when we publish S3 assets and when we publish ECR assets. The reason is that ECR publishers execute `docker build` which can potentially execute 3rd party code (via a base docker image).

This change modifies the conventional name for the publishing roles as well as adds a set of properties to the `DefaultStackSynthesizer` to allow customization as needed.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*