tighten "stage.outdir" and omit redundant "addToParentAssembly"
It should not be possible to specify the output directory for a nested stage, only for apps.

Also, make assemblyBuilder internal. It is only needed from the private "synthesize.ts" module (and the reason it is a separate module is for backwards compat. in an ideal world synthesis will be implemented the same file).

There is no need for "addToParentAssembly" because we already do that when we create the nested assembly.