feat(stepfunctions-tasks): task constructs for creating and transforming SageMaker jobs (#8391)
replacement for the current implementation of `SageMaker` service
integration and state level properties are merged and represented
as a construct.

The previous implementation that implemented `IStepFunctionsTask` has
been removed. The previously existing classes were directly converted to constructs
as these were marked **experimental** and still require further iterations
as they are not backed by a SageMaker L2 (does not exist yet).

In the interest of pragmatism, I decided to move them to leverage the newer pattern
so we can deprecate the `Task` construct.

Note that I have left the unit and integration tests verbatim. The integration test
requires some additional steps as there are pre-requisites to running a training job
such as creating and configuring input data that are not currently included.

BREAKING CHANGE: constructs for `SageMakerCreateTrainingJob` and
`SageMakerCreateTransformJob` replace previous implementation that
implemented `IStepFunctionsTask`.
* **stepfunctions-tasks:** `volumeSizeInGB` property in `ResourceConfig` for
SageMaker tasks are now type `core.Size`
* **stepfunctions-tasks:** `maxPayload` property in `SagemakerTransformProps`
is now type `core.Size`
* **stepfunctions-tasks:** `volumeKmsKeyId` property in `SageMakerCreateTrainingJob` is now `volumeEncryptionKey`

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*