feat(core,s3-assets,lambda): custom asset bundling (#7898)
Adds support for asset bundling by running a command inside a Docker container.

The asset path is mounted in the container at `/asset-input` and is set as the working
directory. The container is responsible for putting content at `/asset-output`. The content
at `/asset-output` will be zipped and used as the final asset.

This allows to use Docker for Lambda code bundling. 

It will also be possible to refactor `aws-lambda-nodejs` and create other language
specific modules.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*