chore: remove awkward cfn2ts script entries (#8231)
Packages that are not containers of L1 libraries (`Cfn~` classes) have
no point in having a `cfn2ts` script registered. This causes problems
when trying to generate L1s across the whole repository using
`lerna run cfn2ts`.

This adds a `pkglint` rule that mandates the `cfn2ts` script is only
present when the related other metadata is also required to be present.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*