fix(eks): fargate profile role is not added to aws-auth by the cdk
When a Fargate Profile is added to the cluster, we need to make sure the aws-auth config map is updated from within the CDK app. EKS will do that behind the scenes if it's not done manually, but this means that it would be an out-of-band update of the config map and will be overridden by the CDK if the config map is updated manually.

Fixes #7981

BREAKING CHANGE: `cluster.awsAuth` can now return `undefined` if the cluster is not kubectl-enabled. This ensures that users take into account the fact that the cluster may not support aws-auth updates. In most cases, it's sufficient to just use this syntax `cluster.awsAuth?.addxxx` to conditionally update aws-auth.