fix(cli): bootstrapping cannot be retried
We used to failing if the bootstrap stack was not either missing or
in a completely successful state.

Specifically, recoverable states like `ROLLBACK_COMPLETE` or
`UPDATE_ROLLBACK_COMPLETE` were treated as failure conditions, even
though the `deployStack()` that would be called later knows how to
recover from those.

Fix the unnecessary pessimism.