fix(cli): bootstrapping cannot be retried (#8577)
We used to failing if the bootstrap stack was not either missing or
in a completely successful state.

Specifically, recoverable states like `ROLLBACK_COMPLETE` or
`UPDATE_ROLLBACK_COMPLETE` were treated as failure conditions, even
though the `deployStack()` that would be called later knows how to
recover from those.

Fix the unnecessary pessimism.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*