chore: option to skip the next directory in foreach.sh (#8477)
When running arbitrary commands using `foreach.sh`, packages may fail
because the commands do not work for good reason. For example, if the
command relies on the use of 'cdk-build' but the package does not
use 'cdk-build', this will fail.

Depending on the intention with which this script was invoked, it may
be acceptable for this package to be skipped, and continue with the
execution.

The `--skip` option offers this functionality where running `foreach.sh
--skip` will simply remove the package from the queue, allowing the
script to continue with the rest of the packages.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*