feat(cognito): user pool client - disable OAuth easily (#8496)
Currently, the defaults for the `UserPoolClient` sets up OAuth flows and
scopes. If a user wishes to not use OAuth, they have to explicitly set
a number of properties to `undefined` and use the escape hatch to
control a property that does not have direct access at the L2 construct.

See attached issue for details on how this is painful.

A new property `disableOAuth` allows a user control this.

closes #8429


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*