feat(cli): new deployment monitoring (#8165)
Implement a new stack monitor, which updates the screen in-place and
shows the resources currently being deployed plus a progress bar,
instead of dumping a giant history of all events to the screen.

The new monitor can is activated automatically if the output is a TTY,
if we're not on Windows, and if we're not doing a verbose run.

Fix in both old and new stack monitors: detect rollback events, and
instead of increasing the event counter (leading to a progress of
`84/45` resources done), roll the counter backwards as resources are
being rolled back.


![anim](https://user-images.githubusercontent.com/524162/83119825-7a141680-a0d0-11ea-8110-4cb80c9f82a1.gif)



----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*