Reduces IAM permissions granted for KMS key.
This changes the permissions appended to the KMS key policy such they
are as minimal as possible. Permissions are split between a grant when
creating the volume -- a grant that allows the Volume to be created --
and a grant when allowing attachment of the Volume.

This properly restricts use of the CMK-encrypted volume to only those
users that have access to both the EBS volume and, separately, the CMK
used for encryption.