feat(logs): MetricFilter exposes extracted Metric object (#8556)
It is convention to expose metrics on AWS constructs, but metric filters currently do not.

This PR adds a `metric()` API that creates a metric from a metric filter with the same name and namespace. The user can specify what statistic to use, or it will default to average over 5 minutes.

fixes #1353 

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*