fix(ec2): can't set natGateways=0 using reserved private subnets (#8407)
Having only reserved private subnets throws an error when setting `natGateways` to `0` in the VPC subnet configuration.
 
Reserved subnets are not provisioned, so it should be possible to do this.

This PR adds a check to make sure that the error is thrown only when `natGateways=0` and private subnets are being provisioned.

fixes #8203 
----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*