chore(stepfunctions-tasks): task state to call Amazon EMR  (#8177)
replacement for the current implementation of `EMR` service integration.
Merges state and service integration properties and represents them as a
construct.

The previously implementation that implemented `IStepFunctionsTask` were
marked **experimental**. There's a fair amount of cleanup work ahead after
this shift over. Prioritized deprecating the current `Task` class.

Opportunistically made some improvements along the way, but it's not
thorough by any means. Further iterations will be needed to leverage the L2
as well as re-organize this code as `create-cluster` is a monster class.

There were a bunch of exported functions in the implementation that are
incompatible with the jsii type model so the declarations were being ignored.
Shifted these into a private folder. They'll require further organization as well.

I have left the unit and integration tests as they used to be.

BREAKING CHANGE: constructs for `EMR*` have been introduced to replace
previous implementation which implemented `IStepFUnctionsTask`.
* **stepfunctions-tasks:** `sizeInGB` property in `VolumeSpecification` has been renamed to `volumeSize` and is of type `cdk.Size` as we want to enable specifying any unit
* **stepfunctions-tasks:** `ebsRootVolumeSize` property in `EmrCreateCluster` is now of type `cdk.Size` as we want to enable specifying any unit
* **stepfunctions-tasks:** `Tags` in `EmrCreateCluster` type has changed from `cdk.CfnTag[]` to a map of string to string as we do not want to leak `Cfn` types


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*