feat(stepfunctions): grant APIs for state machine construct (#8486)
This PR closes #5933. It adds additional grant APIs to supplement `grantStartExecution()`. 

API additions to `state-machine.ts`:

- `grantRead()` method that grants read access to a role.
- `grantTaskResponse()` method that grants task response access to a role.
- `grantExecution()` method that allows user to specify what action to map onto all executions.
- `grant()` method that allows user to specify what action to map onto the state machine.

API additions to `activity.ts`:

- `grant()` method that allows user to specify what action to map onto the activity.


The idea behind these API methods is to mimic the convention of other `grant()` APIs in other modules. This is slightly more difficult with Step Functions because of the multiple resources that IAM actions can map onto. The rationale is that `grant()` in `state-machine.ts` will handle all custom permissions for the state machine, `grantExecution()` in `state-machine.ts` will handle all custom permissions for the executions, and `grant()` in `activity.ts` will handle all custom permissions for activities.

`grantRead()` and `grantTaskResponse()` are convenience APIs that were specified in the original issue for this feature.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*