fix(toolkit): CLI tool fails on CloudFormation Throttling
The CDK (particularly, `cdk deploy`) might crash after getting throttled
by CloudFormation, after the default configured 6 retries has been
reached.

This changes the retry configuration of the CloudFormation client (and
only that one) to allow up to 10 retries with a backoff base of 1
second. This makes the maximum back-off about 17 minutes, which I hope
would be plenty enough even for the 1 TPM calls. This should allow
heavily parallel deployments on the same account and region to avoid
getting killed by a throttle; but will reduce the responsiveness of the
progress UI.

Additionaly, configured a custom logger for the SDK, which would log the
SDK calls to the console when running in debug mode, allowing the users
to gain visibility on more information for troubleshooting purposes.

Fixes #5637