feat(cfn-include): support logical id overrides (#8529)
Previously, we created the Tokens needed for references like `Ref` and `Fn::GetAtt`
without actually referencing the created L1s in the template,
just by their logical IDs.
That's not strictly correct,
because users of `CfnInclude` can call `overrideLogicalId()`
on the CloudFormation elements retrieved from the template,
and we need to make sure to also reflect that in all references to that resource.

Change the code converting from CloudFormation to CDK values to actually retrieve
the appropriate L1 objects when encountering expressions like `Ref` and `Fn::GetAtt`.
This also gives us correctly working cross-stack references for free.

Related to #7375

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*