fix(backup): vault name may exceed 50 characters (#8653)
Remove `AWS::StackName` from the CDK generated vault name, it's already
included in the `uniqueId`.

BREAKING CHANGE: existing vaults that use a generated name will be replaced but
existing recovery points won't be lost. The default vault removal policy is
`RETAIN` and if it was set to `DESTROY` the deployment will fail because
vault with recovery points cannot be deleted.

Fixes #8627


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*