fix(cognito): cannot add multiple route53 targets to the same user pool domain (#8622)
The root cause here is that calling `cloudFrontDomainName` getter on
`UserPoolDomain` results in creating a custom resource with a fixed node
id. This resulted in the error - "There is already a Construct with name
'CloudFrontDomainName' in UserPoolDomain".

Changed the logic around so that the `CustomResource` construct is only
created on first call to the getter, and is reused on subsequent calls.

fixes #8603


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*