chore(stepfunctions): make renderTask internal for the task state and derived classes
`renderTask` has been marked internal and renamed to `_renderTask` to align
with naming convention of internal APIs

Motivation:
users do not directly need to interact or use this API.
It's responsible for rendering all Amazon States Language (ASL) properties
for service integrations to Step Functions.

the `toStateJson()` API is public on the `State` class and can be used
in tests for users that want to look at a task state's configuration.

The partial configuration that comes from `_renderTask()` is included
in the `toStateJson()` API.