feat(stepfunctions-tasks): evaluate expression as a task construct (#8555)
Replacement for the current implementation of experimental Evaluate Expression.
Merges state and service integration properties and represents them as a
construct.

Note that the changes to test are only the reordering of elements.
Re-ran the integration test and verified that it functions as documented.

BREAKING CHANGE:
`EvaluateExpression` is now a construct representing a task state rather than an embedded property called `task`

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*