feat(stepfunctions-tasks): task construct to call `RunJob` on ECS (#8451)
Replacement for the current implementation of `ECS` service integration.
Merges state and service integration properties and represents them as a
construct.

The notable differences from the current implementation are:
* single implementation class`runTask` with an added `launchTarget` required
property where the properties specific to the launch type are specified
* linter disable of `ref-via-interface` as we need to use `TaskDefinition`
because properties that are not known for imported task definitions are required
* older implementations have been marked deprecated with direction to the
replacement.

Left all of the unit tests (except fargate platform version which was not
previously honored) and integ test expectations verbatim to ensure that
we have not lost fidelity. 

BREAKING CHANGE:
* `containerName` is not supported as an override anymore and has been replaced by `containerDefinition`

Closes #8610


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*