feat(secretsmanager): add grantUpdate method (#8600)
This creates the `grantUpdate` method to give an `iam.IGrantable` the ability to update a secret in Secrets Manager.

I opted to move it out from the `grantWrite` method, as there may be cases where we want something to write an initial value, but not change it later on. This decision could be reversed if it seems too granular or too much of an edge-case.

Closes #8491 

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*