chore(stepfunctions): minor updates to README (#8823)
removed `Typescript` prefix ahead of example as we translate code snippets
for each of our supported languages in the module overview.

dropped the `Future Work` section since 2 of these items have been completed
and the other is being tracked in GitHub issues. This aligns with other `README`
sections for our modules.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*